Avoid requiring user's server to be Clone
The old code took a bit of a shortcut, cloning the user's Server for
every request. This was hidden behind tokio_proto::NewService, but if
you look at it, all it does is clone its argument whenever new_service()
is invoked. This is inconvenient for users, as it means they need to
carefully craft their Server such that it is in fact Clone. This will
likely involve adding reference counts to all fields, or create an Arc
wrapper, implement Service for it, and then using that to call
.listen().

This commit essentially does this for the user. The Server given by the
user is wrapped in an Arc that implements Service if the contained value
does. When a new request arrives, only this Arc is cloned, not the
underlying Server. Since all the methods on Service take & receivers,
this works fine. It should also lead to a decent performance improvement
as we do potentially much less work for every incoming request.

In terms of API changes, this relaxes the requirements on the user's
Server type from

  Send + 'static + Clone (which adds + Sized)

to

  Send + 'static + Sync + Sized

which should be backwards compatible for all but the weirdest types
(i.e., something which implements Clone but not Sync). If we want to
support those too, this is doable, but we should put off that added
complexity until we have a use-case for it.